BUG FIX: wrong attribute/label selection when building filter for inverse rule
When finding inverse rule, if there is a set of labels with some items
with the same initial characters (i.e attribute=0 and attribute=0.1) the
first occurrence will be selected ever. With this fix the algorithm
looks for the label's end that must be an empty space or the end of
string.